stress-chmod: fix incorrect type in 3rd arg in do_fchmod
The index argument i should be size_t and not int. Fix it.

+++ stress-chmod.c
@@ -91,7 +91,7 @@ static const mode_t modes[] = {
 static int do_fchmod(
 	const int fd,
 	const int bad_fd,
-	const int i,
+	const size_t i,
 	const mode_t mask,
 	const mode_t all_mask)
 {
